\name{a.Crowder.Mod}
\alias{a.Crowder.Mod}
\title{First weighting vector of the modified quadratic estimating equation (Crowder)}
\usage{
a.Crowder.Mod(param, Y, variancef, dmean, dsd)

a.Crowder.Mod(param, Y, variancef, dmean, dsd)
}
\arguments{
  \item{param}{Vector of parameters of the distribution
  function}

  \item{Y}{Individual data sample}

  \item{variancef}{Variance function of the distribution}

  \item{dmean}{Derivative in respect to the parameter
  vector of the mean function of the distribution}

  \item{dsd}{Derivative in respect to the parameter vector
  of the standard deviation function of the distribution}

  \item{param}{Vector of parameters of the distribution
  function}

  \item{Y}{Individual data sample}

  \item{variancef}{Variance function of the distribution}

  \item{dmean}{Derivative in respect to the parameter
  vector of the mean function of the distribution}

  \item{dsd}{Derivative in respect to the parameter vector
  of the standard deviation function of the distribution}
}
\value{
First weighting vector

First weighting vector
}
\description{
First weighting vector of the modified quadratic estimating
equation (Crowder)

Second weighting vector of the modified quadratic
estimating equation (Crowder)
}
\author{
Francois Pelletier

Francois Pelletier
}

