\name{a.Crowder}
\alias{a.Crowder}
\title{First weighting vector of the quadratic estimating equation (Crowder)}
\usage{
a.Crowder(param, variancef, skewnessf, kurtosisf, dmean, dsd)
}
\arguments{
  \item{param}{Vector of parameters of the distribution
  function}

  \item{variancef}{Variance function of the distribution}

  \item{skewnessf}{Skewness function of the distribution}

  \item{kurtosisf}{Kurtosis function of the distribution}

  \item{dmean}{Derivative in respect to the parameter
  vector of the mean function of the distribution}

  \item{dsd}{Derivative in respect to the parameter vector
  of the standard deviation function of the distribution}
}
\value{
First weighting vector
}
\description{
First weighting vector of the quadratic estimating equation
(Crowder)
}
\author{
Francois Pelletier
}

