\name{a.gauss}
\alias{a.gauss}
\title{First weighting vector of the quadratic estimating equation (gaussian)}
\usage{
a.gauss(param, variancef, dmean)
}
\arguments{
  \item{param}{Vector of parameters of the distribution
  function}

  \item{variancef}{Variance function of the distribution}

  \item{dmean}{Derivative in respect to the parameter
  vector of the mean function of the distribution}
}
\value{
First weighting vector
}
\description{
First weighting vector of the quadratic estimating equation
(gaussian)
}
\author{
Francois Pelletier
}

