\name{b.Crowder.Mod}
\alias{b.Crowder.Mod}
\title{Second weighting vector of the modified quadratic estimating equation (Crowder)}
\usage{
b.Crowder.Mod(param, Y, variancef, dmean, dsd)
}
\arguments{
  \item{param}{Vector of parameters of the distribution
  function}

  \item{Y}{Individual data sample}

  \item{variancef}{Variance function of the distribution}

  \item{dmean}{Derivative in respect to the parameter
  vector of the mean function of the distribution}

  \item{dsd}{Derivative in respect to the parameter vector
  of the standard deviation function of the distribution}
}
\value{
First weighting vector
}
\description{
Second weighting vector of the modified quadratic
estimating equation (Crowder)
}
\author{
Francois Pelletier
}

