\name{b.gauss}
\alias{b.gauss}
\title{Second weighting vector of the quadratic estimating equation (gaussian)}
\usage{
b.gauss(param, variancef, dsd)
}
\arguments{
  \item{param}{Vector of parameters of the distribution
  function}

  \item{variancef}{Variance function of the distribution}

  \item{dsd}{Derivative in respect to the parameter vector
  of the standard deviation function of the distribution}
}
\value{
Second weighting vector
}
\description{
Second weighting vector of the quadratic estimating
equation (gaussian)
}
\author{
Francois Pelletier
}

