\name{covariance.QEE}
\alias{covariance.QEE}
\title{Covariance matrix using the component matrix M and V from QEE}
\usage{
covariance.QEE(M, V, n)
}
\arguments{
  \item{M}{Gradient matrix}

  \item{V}{Covariance matrix of equations}

  \item{n}{Sample size}
}
\value{
Weighted covariance matrix
}
\description{
Covariance matrix using the component matrix M and V from
QEE
}
\author{
François Pelletier
}

