\name{gammaf.Crowder.Mod}
\alias{gammaf.Crowder.Mod}
\title{Gamma function used in Modified Crowder Estimating Equations}
\usage{
gammaf.Crowder.Mod(Y)
}
\arguments{
  \item{Y}{Individual data sample}
}
\value{
Gamma function value
}
\description{
Gamma function used in Modified Crowder Estimating
Equations
}
\author{
Francois Pelletier
}

