\name{gammaf.Crowder}
\alias{gammaf.Crowder}
\title{Gamma function used in Crowder Estimating Equations}
\usage{
gammaf.Crowder(param, skewnessf, kurtosisf)
}
\arguments{
  \item{param}{Vector of parameters of the distribution
  function}

  \item{skewnessf}{Skewness function of the distribution}

  \item{kurtosisf}{Kurtosis function of the distribution}
}
\value{
Gamma function value
}
\description{
Gamma function used in Crowder Estimating Equations
}
\author{
Francois Pelletier
}

