\name{M.Crowder.Mod}
\alias{M.Crowder.Mod}
\title{M Matrix (Modified Crowder)}
\usage{
M.Crowder.Mod(param, Y, variancef, skewnessf, kurtosisf, dmean, dsd)
}
\arguments{
  \item{Y}{Individual data sample}

  \item{param}{Vector of parameters of the distribution
  function}

  \item{variancef}{Variance function of the distribution}

  \item{skewnessf}{Skewness function of the distribution}

  \item{kurtosisf}{Kurtosis function of the distribution}

  \item{dmean}{Derivative in respect to the parameter
  vector of the mean function of the distribution}

  \item{dsd}{Derivative in respect to the parameter vector
  of the standard deviation function of the distribution}
}
\value{
M Matrix
}
\description{
M Matrix (Modified Crowder)
}
\author{
Francois Pelletier
}

