\name{V.Crowder.Mod}
\alias{V.Crowder.Mod}
\title{V Matrix (Modified Crowder)}
\usage{
V.Crowder.Mod(param, Y, variancef, dmean, dsd)
}
\arguments{
  \item{Y}{Individual data sample}

  \item{param}{Vector of parameters of the distribution
  function}

  \item{variancef}{Variance function of the distribution}

  \item{dmean}{Derivative in respect to the parameter
  vector of the mean function of the distribution}

  \item{dsd}{Derivative in respect to the parameter vector
  of the standard deviation function of the distribution}
}
\value{
V Matrix
}
\description{
V Matrix (Modified Crowder)
}
\author{
Francois Pelletier
}

