\name{V.Crowder}
\alias{V.Crowder}
\title{V Matrix (Crowder)}
\usage{
V.Crowder(param, Y, variancef, skewnessf, kurtosisf, dmean, dsd)
}
\arguments{
  \item{Y}{Individual data sample}

  \item{param}{Vector of parameters of the distribution
  function}

  \item{variancef}{Variance function of the distribution}

  \item{skewnessf}{Skewness function of the distribution}

  \item{kurtosisf}{Kurtosis function of the distribution}

  \item{dmean}{Derivative in respect to the parameter
  vector of the mean function of the distribution}

  \item{dsd}{Derivative in respect to the parameter vector
  of the standard deviation function of the distribution}
}
\value{
V Matrix
}
\description{
V Matrix (Crowder)
}
\author{
Francois Pelletier
}

